Development and Implementation of Layout
Through a trial and error process I continued to iterate on the original layout that I had tried.
I settled on a vertical and horizontal layout of the same style - boxes with rounded corners.
These layouts are subject to chance but currently will display the data in a clear and concise manner.