Enable Button Control
Enable a Play/Pause and reset button for the ride telemetry. Functions like duration timer, average speed, and total distance can be started, stopped, and reset by user input.

The issue remains that once the timer has been stopped it cannot be started again, this will be fixed in the future.